fix the renaming of ProblematicDiaryEntry to DiaryEntry
