+++ notifyme-sdk-backend/notifyme-sdk-backend-data/src/test/java/ch/ubique/notifyme/sdk/backend/data/DataServiceTestHelper.java
@@ -10,7 +10,7 @@
 
 package ch.ubique.notifyme.sdk.backend.data;
 
-import ch.ubique.notifyme.sdk.backend.model.ProblematicDiaryEntryWrapper.DiaryEntry;
+import ch.ubique.notifyme.sdk.backend.model.DiaryEntryWrapperOuterClass.DiaryEntry;
 import ch.ubique.notifyme.sdk.backend.model.VenueTypeOuterClass.VenueType;
 import ch.ubique.notifyme.sdk.backend.model.event.JavaDiaryEntry;
 import java.util.Calendar;
