+++ notifyme-sdk-backend/notifyme-sdk-backend-model/src/main/resources/proto/DiaryEntryWrapper.proto
