+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/main/java/ch/ubique/notifyme/sdk/backend/ws/controller/NotifyMeController.java
@@ -10,12 +10,18 @@
 
 package ch.ubique.notifyme.sdk.backend.ws.controller;
 
+import ch.ubique.notifyme.sdk.backend.data.NotifyMeDataService;
+import ch.ubique.notifyme.sdk.backend.model.ProblematicEventWrapperOuterClass.ProblematicEvent;
+import ch.ubique.notifyme.sdk.backend.model.ProblematicEventWrapperOuterClass.ProblematicEvent.Builder;
+import ch.ubique.notifyme.sdk.backend.model.ProblematicEventWrapperOuterClass.ProblematicEventWrapper;
+import ch.ubique.notifyme.sdk.backend.model.tracekey.TraceKey;
+import ch.ubique.notifyme.sdk.backend.model.util.DateUtil;
+import ch.ubique.openapi.docannotations.Documentation;
+import com.google.protobuf.ByteString;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
-
 import javax.validation.Valid;
-
 import org.springframework.http.CacheControl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -28,19 +34,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.google.protobuf.ByteString;
-
-import ch.ubique.notifyme.sdk.backend.data.NotifyMeDataService;
-import ch.ubique.notifyme.sdk.backend.model.ProblematicEventWrapperOuterClass.ProblematicEvent;
-import ch.ubique.notifyme.sdk.backend.model.ProblematicEventWrapperOuterClass.ProblematicEvent.Builder;
-import ch.ubique.notifyme.sdk.backend.model.ProblematicEventWrapperOuterClass.ProblematicEventWrapper;
-import ch.ubique.notifyme.sdk.backend.model.tracekey.TraceKey;
-import ch.ubique.notifyme.sdk.backend.model.util.DateUtil;
-import ch.ubique.openapi.docannotations.Documentation;
-
 @Controller
 @RequestMapping("/v1")
-@CrossOrigin(origins = { "https://notify-me.c4dt.org", "https://notify-me-dev.c4dt.org"})
+@CrossOrigin(origins = {"https://notify-me.c4dt.org", "https://notify-me-dev.c4dt.org"})
 public class NotifyMeController {
     private static final String HEADER_X_KEY_BUNDLE_TAG = "x-key-bundle-tag";
 
@@ -50,7 +46,10 @@ public class NotifyMeController {
     private final Long traceKeysCacheControlInMs;
 
     public NotifyMeController(
-            NotifyMeDataService dataService, String revision, Long bucketSizeInMs, Long traceKeysCacheControlInMs) {
+            NotifyMeDataService dataService,
+            String revision,
+            Long bucketSizeInMs,
+            Long traceKeysCacheControlInMs) {
         this.dataService = dataService;
         this.revision = revision;
         this.bucketSizeInMs = bucketSizeInMs;
@@ -76,7 +75,7 @@ public class NotifyMeController {
             return ResponseEntity.notFound().build();
         }
         return ResponseEntity.ok()
-                .cacheControl(CacheControl.maxAge(traceKeysCacheControlInMs, TimeUnit.MILLISECONDS))              
+                .cacheControl(CacheControl.maxAge(traceKeysCacheControlInMs, TimeUnit.MILLISECONDS))
                 .header(
                         HEADER_X_KEY_BUNDLE_TAG,
                         Long.toString(DateUtil.getLastFullBucketEndEpochMilli(bucketSizeInMs)))
@@ -109,7 +108,10 @@ public class NotifyMeController {
                                     "in millis since epoch. must be aligned to a full hour, and < now()")
                     Long lastKeyBundleTag) {
         if (!isValidKeyBundleTag(lastKeyBundleTag)) {
-            return ResponseEntity.notFound().cacheControl(CacheControl.maxAge(traceKeysCacheControlInMs, TimeUnit.MILLISECONDS)).build();
+            return ResponseEntity.notFound()
+                    .cacheControl(
+                            CacheControl.maxAge(traceKeysCacheControlInMs, TimeUnit.MILLISECONDS))
+                    .build();
         }
         List<TraceKey> traceKeys = dataService.findTraceKeys(DateUtil.toInstant(lastKeyBundleTag));
         ProblematicEventWrapper pew =
@@ -118,7 +120,7 @@ public class NotifyMeController {
                         .addAllEvents(mapToProblematicEvents(traceKeys))
                         .build();
         return ResponseEntity.ok()
-                .cacheControl(CacheControl.maxAge(traceKeysCacheControlInMs, TimeUnit.MILLISECONDS))       
+                .cacheControl(CacheControl.maxAge(traceKeysCacheControlInMs, TimeUnit.MILLISECONDS))
                 .header("content-type", "application/x-protobuf")
                 .header(
                         HEADER_X_KEY_BUNDLE_TAG,
@@ -127,8 +129,7 @@ public class NotifyMeController {
     }
 
     private List<ProblematicEvent> mapToProblematicEvents(List<TraceKey> traceKeys) {
-        return traceKeys.stream()
-                .map(t -> mapTraceKeyToProblematicEvent(t))
+        return traceKeys.stream().map(this::mapTraceKeyToProblematicEvent)
                 .collect(Collectors.toList());
     }
