+++ notifyme-sdk-backend/notifyme-sdk-backend-ws/src/test/java/ch/ubique/notifyme/sdk/backend/ws/controller/web/WebCriticalEventControllerTest.java
@@ -5,7 +5,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import ch.ubique.notifyme.sdk.backend.model.ProblematicDiaryEntryWrapper.DiaryEntryWrapper;
 import ch.ubique.notifyme.sdk.backend.ws.controller.BaseControllerTest;
 import ch.ubique.notifyme.sdk.backend.ws.controller.DebugControllerTestHelper;
 import org.hamcrest.Matchers;
@@ -19,7 +18,7 @@ public class WebCriticalEventControllerTest extends BaseControllerTest {
     public void setUp() throws Exception {
         criticalEventEndPoint = "/criticalevent";
 
-        final DiaryEntryWrapper wrapper = DebugControllerTestHelper.getTestDiaryEntryWrapper();
+        final var wrapper = DebugControllerTestHelper.getTestDiaryEntryWrapper();
 
         mockMvc.perform(
                 post("/v1/debug/criticalevent")
@@ -29,7 +28,7 @@ public class WebCriticalEventControllerTest extends BaseControllerTest {
 
     @Test
     public void shouldReturnHtml() throws Exception {
-        final DiaryEntryWrapper wrapper = DebugControllerTestHelper.getTestDiaryEntryWrapper();
+        final var wrapper = DebugControllerTestHelper.getTestDiaryEntryWrapper();
 
         mockMvc.perform(
                         get(criticalEventEndPoint)
